requirements.txt
torch_scatter
torch_geometric
pytorch-lightning
torch_cluster
torch-sparse